Move `expandConfigRequest` method to config.js to avoid dependency.
